Enforce strict casts rule
The `strict-casts` analyzer option is enabled on `flutter/website` sample code and possibly makes sense to enable here as well. It is the modern version of the old “implicit-casts” strong mode option.
https://dart.dev/tools/analysis#enabling-additional-type-checks

+++ samples/multiplayer/lib/game_internals/playing_card.dart
@@ -18,7 +18,7 @@ class PlayingCard {
     return PlayingCard(
       CardSuit.values
           .singleWhere((e) => e.internalRepresentation == json['suit']),
-      json['value'],
+      json['value'] as int,
     );
   }
